refactor(node): extract signer and payload modules from main (#2474)
